gtfs: move realtime init code to where the rest is
