+++ web-bundle/src/main/java/com/graphhopper/http/GraphHopperBundleConfiguration.java
@@ -24,4 +24,6 @@ public interface GraphHopperBundleConfiguration {
 
     GraphHopperConfig getGraphHopperConfiguration();
 
+    RealtimeConfiguration gtfsrealtime();
+
 }
