+++ web-bundle/src/main/java/com/graphhopper/http/RealtimeFeedLoadingCache.java
@@ -25,11 +25,9 @@ import com.google.common.cache.LoadingCache;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListenableFutureTask;
 import com.google.transit.realtime.GtfsRealtime;
-import com.graphhopper.gtfs.GtfsStorage;
+import com.graphhopper.gtfs.GraphHopperGtfs;
 import com.graphhopper.gtfs.RealtimeFeed;
 import com.graphhopper.gtfs.Transfers;
-import com.graphhopper.routing.util.EncodingManager;
-import com.graphhopper.storage.BaseGraph;
 import io.dropwizard.lifecycle.Managed;
 import org.apache.hc.client5.http.classic.HttpClient;
 import org.apache.hc.client5.http.classic.methods.HttpGet;
@@ -48,27 +46,24 @@ import java.util.concurrent.TimeUnit;
 public class RealtimeFeedLoadingCache implements Factory<RealtimeFeed>, Managed {
 
     private final HttpClient httpClient;
-    private final BaseGraph baseGraph;
-    private final EncodingManager encodingManager;
-    private final GtfsStorage gtfsStorage;
-    private final RealtimeBundleConfiguration bundleConfiguration;
+    private final GraphHopperGtfs graphHopper;
+    private final GraphHopperBundleConfiguration bundleConfiguration;
     private ExecutorService executor;
     private LoadingCache<String, RealtimeFeed> cache;
     private Map<String, Transfers> transfers;
 
     @Inject
-    RealtimeFeedLoadingCache(BaseGraph baseGraph, EncodingManager encodingManager, GtfsStorage gtfsStorage, HttpClient httpClient, RealtimeBundleConfiguration bundleConfiguration) {
-        this.baseGraph = baseGraph;
-        this.encodingManager = encodingManager;
-        this.gtfsStorage = gtfsStorage;
+    RealtimeFeedLoadingCache(GraphHopperGtfs graphHopper, HttpClient httpClient, GraphHopperBundleConfiguration bundleConfiguration) {
+        this.graphHopper = graphHopper;
         this.bundleConfiguration = bundleConfiguration;
         this.httpClient = httpClient;
     }
 
     @Override
     public void start() {
+        System.out.println("Starting RealtimeFeedLoadingCache");
         this.transfers = new HashMap<>();
-        for (Map.Entry<String, GTFSFeed> entry : this.gtfsStorage.getGtfsFeeds().entrySet()) {
+        for (Map.Entry<String, GTFSFeed> entry : this.graphHopper.getGtfsStorage().getGtfsFeeds().entrySet()) {
             this.transfers.put(entry.getKey(), new Transfers(entry.getValue()));
         }
         this.executor = Executors.newSingleThreadExecutor();
@@ -112,14 +107,24 @@ public class RealtimeFeedLoadingCache implements Factory<RealtimeFeed>, Managed 
         Map<String, GtfsRealtime.FeedMessage> feedMessageMap = new HashMap<>();
         for (FeedConfiguration configuration : bundleConfiguration.gtfsrealtime().getFeeds()) {
             try {
-                GtfsRealtime.FeedMessage feedMessage = httpClient.execute(new HttpGet(configuration.getUrl().toURI()),
-                        response -> GtfsRealtime.FeedMessage.parseFrom(response.getEntity().getContent()));
-                feedMessageMap.put(configuration.getFeedId(), feedMessage);
+                switch (configuration.getUrl().getProtocol()) {
+                    case "http": {
+                        GtfsRealtime.FeedMessage feedMessage = httpClient.execute(new HttpGet(configuration.getUrl().toURI()),
+                                response -> GtfsRealtime.FeedMessage.parseFrom(response.getEntity().getContent()));
+                        feedMessageMap.put(configuration.getFeedId(), feedMessage);
+                        break;
+                    }
+                    case "file": {
+                        GtfsRealtime.FeedMessage feedMessage = GtfsRealtime.FeedMessage.parseFrom(configuration.getUrl().openStream());
+                        feedMessageMap.put(configuration.getFeedId(), feedMessage);
+                        break;
+                    }
+                }
             } catch (IOException | URISyntaxException e) {
                 throw new RuntimeException(e);
             }
         }
-        return RealtimeFeed.fromProtobuf(gtfsStorage, this.transfers, feedMessageMap);
+        return RealtimeFeed.fromProtobuf(graphHopper.getGtfsStorage(), this.transfers, feedMessageMap);
     }
 
 }
